requirements.txt
Django==4.2.7
django-extensions==3.2.3
django-ninja==1.0.1
pytest-django==4.7.0
python-decouple==3.8
django-cors-headers==4.3.1
django-ninja-jwt==5.2.10
